+++ test/src/test_mat3x2.h
@@ -142,7 +142,7 @@ TEST_IMPL(GLM_PREFIX, mat3x2_scale) {
   int i, j;
   float scale;
 
-  scale = rand() % 100;
+  scale = (float)(rand() % 100);
 
   GLM(mat3x2_scale)(m1, scale);
